changed around folder structure to make more sense, and changed up some naming conventions.
