+++ client/src/App.css
@@ -1,3 +1,7 @@
+footer {
+    text-align: center;
+}
+
 .legacyTitle {
     font-family: 'IM Fell English', serif;
 }
@@ -14,4 +18,4 @@
     display: flex;
     justify-content: center;
     align-items: center;
- }
+}
