--- client/src/components/LandingPage.js
@@ -1,46 +1,0 @@
-import React from 'react';
-import Container from 'react-bootstrap/Container';
-import Row from 'react-bootstrap/Row';
-import Col from 'react-bootstrap/Col';
-
-function LandingPage() {
-  
-  return (
-    <main>
-        <Container>
-            <Row>
-                <Col>
-                    <div className="landingPageLogoWrappingText">Welcome to</div>
-                </Col>
-            </Row>
-            <Row>
-                <Col>
-                    <h1 className="headingFont">Legacy</h1>
-                </Col>
-            </Row>
-            <Row>
-                <Col>
-                    <div className="landingPageLogoWrappingText">The app for people with Dimentia and their loved ones to remember their life memories.</div>
-                </Col>
-            </Row>
-            <Row>
-                <Col>
-                    <img src="./../media/legacyLogo.png" className="landingPageLogo" alt="Legacy Logo"/>
-                </Col>
-            </Row>
-            <Row>
-                <Col>
-                    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
-                </Col>
-            </Row>
-            <Row>
-                <Col>
-                    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
-                </Col>
-            </Row>
-        </Container>
-    </main>
-  );
-}
-
-export default LandingPage;
