--- client/src/components/LogoNav/index.js
@@ -1,3 +1,0 @@
-import LogoNav from "./LogoNav";
-
-export default LogoNav;
