--- client/src/components/SiteNav/index.js
@@ -1,3 +1,0 @@
-import SiteNav from "./SiteNav";
-
-export default SiteNav;
