+++ client/src/components/footer/Footer.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import Container from 'react-bootstrap/Container';
+import Col from 'react-bootstrap/Col';
+import Row from 'react-bootstrap/Row';
+
+function Footer() {
+    return (
+        <footer>
+            <Container>
+                <Col>
+                    <Row>
+                        <p>Copyright©2022</p>
+                    </Row>
+                </Col>
+            </Container>
+        </footer>
+    )
+}
+
+export default Footer;
