+++ client/src/components/header/Header.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import SiteNav from './SiteNav';
+import LogoNav from './LogoNav';
+
+function Header() {
+    return (
+        <header>   
+            <LogoNav/>
+            <SiteNav/>
+        </header>
+    )
+}
+
+export default Header;
